Expand int size of IDs; add file content test.

+++ rest-client/src/main/java/com/fwmotion/threescale/cms/model/CmsFile.java
@@ -11,8 +11,8 @@ public class CmsFile implements CmsObject {
 
     private OffsetDateTime createdAt;
     private OffsetDateTime updatedAt;
-    private Integer id;
-    private Integer sectionId;
+    private Long id;
+    private Long sectionId;
     private String path;
     private Set<String> tags;
     private Boolean downloadable;
@@ -41,19 +41,19 @@ public class CmsFile implements CmsObject {
     }
 
     @Override
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getSectionId() {
+    public Long getSectionId() {
         return sectionId;
     }
 
-    public void setSectionId(Integer sectionId) {
+    public void setSectionId(Long sectionId) {
         this.sectionId = sectionId;
     }
